TxProof for AvidM
