+++ types/src/v0/v0_3/txproof.rs
@@ -1,0 +1,8 @@
+use super::{AvidMNsProof, TxIndex};
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+pub struct AvidMTxProof {
+    pub(crate) tx_index: TxIndex,
+    pub(crate) ns_proof: AvidMNsProof,
+}
